added force simulation to the app. added a heirarchy of nodes.

+++ src/Model/ChunkTranslate.java
@@ -6,5 +6,5 @@ public interface ChunkTranslate {
 
     void setup(int x_translate, int y_translate, int x_scale, int y_scale);
 
-    Point translate(int x, int y);
+    Point translate(double x, double y);
 }
